Fixed bug causing infinite looping
Fixed bug where is the user's goal get overshot (e.g. Goal was level 19
but the overflow of exp reaches the next level), the code will
enter an infinite loop as it only looked for when the current level
matches the goal and did not account for when it is greater.