Added calculations to character exp calculator
Lots of overlapse between character_exp_calculator and
weapon_exp_calculator and so there are some improvements made to both
files.

Even with the overlapse, I felt there was enough different to
justify both codes to not share common codes. Still
considering how to refactor in a way that exposes less commmon code.